+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## [Released] - 0.1.0
+- Published to hex.pm: https://hex.pm/packages/lexbor_erl
 
 ## [Unreleased] - Improve document serialization to preserve DOCTYPE